% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl.R
\name{afl}
\alias{afl}
\title{Create an AFL expression}
\usage{
afl(...)
}
\arguments{
\item{...}{AFL expression argument list dynamically generated by \code{update.afl}, but see the note below}
}
\value{
a \code{\link{scidb}} object
}
\description{
Create an AFL expression
}
\note{
The call list expected to look like: \code{(afl arguments, ...)}
}
\keyword{function}
\keyword{internal}

