% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl.R
\name{aflhelp}
\alias{aflhelp}
\title{Display SciDB AFL operator documentation}
\usage{
aflhelp(topic, db)
}
\arguments{
\item{topic}{an \code{\link{afl}} object from a SciDB database connection, or optionally a character string name}

\item{db}{optional database connection from \code{\link{scidbconnect}} (only needed when \code{topic} is a character string)}
}
\value{
displays help
}
\description{
Display SciDB AFL operator documentation
}
\examples{
\dontrun{
d <- scidbconnect()
aflhelp("list", d)     # explicitly look up a character string
help(d$list)        # same thing via R's \\code{help} function
}
}

