% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{at_least}
\alias{at_least}
\title{Returns TRUE if version string x is greater than or equal to than version y}
\usage{
at_least(x, y)
}
\arguments{
\item{x}{version string like "12.1", "15.12", etc. (non-numeric ignored)}

\item{y}{version string like "12.1", "15.12", etc. (non-numeric ignored)}
}
\value{
logical TRUE if x is greater than or equal to y
}
\description{
Returns TRUE if version string x is greater than or equal to than version y
}

