% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{digest_auth}
\alias{digest_auth}
\title{Convenience function for digest authentication.}
\usage{
digest_auth(db, method, uri, realm = "", nonce = "123456")
}
\arguments{
\item{db}{a scidb database connection object}

\item{method}{digest method}

\item{uri}{uri}

\item{realm}{realm}

\item{nonce}{nonce}
}
\description{
Convenience function for digest authentication.
}
\keyword{internal}
