% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{.attsplitter}
\alias{.attsplitter}
\title{Internal function for processing SciDB attribute schema}
\usage{
.attsplitter(x)
}
\arguments{
\item{x}{a scidb object or schema string}
}
\value{
a data frame with parsed attribute data
}
\description{
Internal function for processing SciDB attribute schema
}
