% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{.dimsplitter}
\alias{.dimsplitter}
\title{Internal function for processing SciDB dimension schema}
\usage{
.dimsplitter(x)
}
\arguments{
\item{x}{a scidb object or schema string}
}
\value{
a data frame with parsed dimension data
}
\description{
Internal function for processing SciDB dimension schema
}
