% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getpwd}
\alias{getpwd}
\title{Simple utility to interactively enter a password without showing it on the screen}
\usage{
getpwd(prompt = "Password:")
}
\arguments{
\item{prompt}{a text prompt to display, defaults to "Password:"}
}
\description{
Simple utility to interactively enter a password without showing it on the screen
}

