% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{\%as\%}
\alias{\%as\%}
\title{AFL array aliasing}
\usage{
x \%as\% y
}
\arguments{
\item{x}{an object of class \code{\link{scidb}} (a scidb array or expression)}

\item{y}{alias name}
}
\description{
AFL array aliasing
}

