% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{help,operator-method}
\alias{help,operator-method}
\title{AFL operator help}
\usage{
\S4method{help}{operator}(topic)
}
\arguments{
\item{topic}{\code{afl} operator}
}
\value{
help summary
}
\description{
AFL operator help
}

