\name{is.temp}
\alias{is.temp}
\title{
is.temp
}
\description{
Is \code{name} a temporary SciDB array?
}
\usage{
is.temp(name)
}
\arguments{
  \item{name}{Either a \code{scidb} or \code{scidbdf} object, or a character
  string representing a stored SciDB array name.}
}
\value{
Returns TRUE if \code{x} is a temporary SciDB array.
}
\author{
B. W. Lewis<blewis@paradigm4.com>
}
