\name{length.scidb}
\alias{length.scidb}
\title{
length
}
\description{
Return the total number of array elements of a SciDB array object.
}
\usage{
\method{length}{scidb}(x)
}
\arguments{
  \item{x}{
A \code{scidb} object.
}
}
\value{
Numeric length of the object in elements.
}
\author{
B. W. Lewis <blewis@paradigm4.com>
}
