% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{ls,afl-method}
\alias{ls,afl-method}
\title{List contents of a SciDB database}
\usage{
\S4method{ls}{afl}(name)
}
\arguments{
\item{name}{\code{afl} SciDB connection object from \code{\link{scidbconnect}}}
}
\value{
a \code{data.frame} listing the contents of the database
}
\description{
List contents of a SciDB database
}

