\name{nrow.scidb}
\alias{nrow.scidb}
\alias{ncol.scidb}
\title{
Number of rows or columns of a SciDB matrix.
}
\description{
Return the number of rows of a SciDB matrix (2-D array).
}
\usage{
\method{nrow}{scidb}(x)
\method{ncol}{scidb}(x)
}
\arguments{
  \item{x}{
A \code{scidb} object.
}
}
\value{
The number of rows of the matrix.
}
\author{
B. W. Lewis <blewis@paradigm4.com>
}
