% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{operators}
\alias{operators}
\title{Base SciDB operators}
\format{A data frame with 4 variables, name, signature, help.}
\source{
Paradigm4
\url{http://paradigm4.com}
}
\usage{
data(operators)
}
\description{
Base SciDB operators as of SciDB version 16.9
}
\keyword{datasets}

