\name{Ops.scidb}
\alias{Ops.scidb}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Binary scidb object operations.
}
\description{
Binary scidb object operations.
}
\usage{
\method{Ops}{scidb}(e1, e2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{e1}{
A scidb matrix reference or double-precision R matrix.
}
  \item{e2}{
A scidb matrix reference or double-precision R matrix.
}
}
\value{
A scidb matrix reference.
}
\author{
B. W. Lewis <blewis@paradigm4.com>
}
