% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print,afl-method}
\alias{print,afl-method}
\title{Print a summary of a \code{afl} SciDB database connection object}
\usage{
\S4method{print}{afl}(x)
}
\arguments{
\item{x}{\code{afl} object}
}
\value{
printed object summary
}
\description{
Print a summary of a \code{afl} SciDB database connection object
}
