% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print,scidb-method}
\alias{print,scidb-method}
\title{Print a summary of a \code{scidb} object}
\usage{
\S4method{print}{scidb}(x)
}
\arguments{
\item{x}{a \code{scidb} object}
}
\value{
printed object summary
}
\description{
Print a summary of a \code{scidb} object
}
