% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.afl}
\alias{print.afl}
\title{Print a summary of a \code{afl} SciDB database connection object}
\usage{
\method{print}{afl}(x, ...)
}
\arguments{
\item{x}{\code{afl} object}

\item{...}{optional arguments (not used)}
}
\value{
printed object summary
}
\description{
Print a summary of a \code{afl} SciDB database connection object
}
