\name{print.scidb}
\alias{print.scidb}
\title{
Print a summary of a SciDB array object.
}
\description{
Print a summary of a SciDB array object.
}
\usage{
\method{print}{scidb}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ A scidb array object.
}
  \item{\dots}{further arguments passed to or from other methods.
}
}
\value{
Printed summary output.
}
\author{
B. W. Lewis <blewis@paradim4.com>
}
