% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl.R
\name{rsub}
\alias{rsub}
\title{Substitute scalar-valued R expressions into an AFL expression
R expressions are marked with R(expression)}
\usage{
rsub(x, env)
}
\arguments{
\item{x}{character valued AFL expression}

\item{env}{environment in which to evaluate R expressions}
}
\value{
character valued AFL expression
}
\description{
Substitute scalar-valued R expressions into an AFL expression
R expressions are marked with R(expression)
}
\keyword{function}
\keyword{internal}
