% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{scidb_unpack_to_dataframe}
\alias{scidb_unpack_to_dataframe}
\title{Unpack and return a SciDB query expression as a data frame}
\usage{
scidb_unpack_to_dataframe(db, query, ...)
}
\arguments{
\item{db}{scidb database connection object}

\item{query}{A SciDB query expression or scidb object}

\item{...}{optional extra arguments (see below)}
}
\description{
Unpack and return a SciDB query expression as a data frame
}
\note{
option extra arguments
\itemize{
  \item{binary}{ optional logical value, if \code{FALSE} use iquery text transfer, otherwise binary transfer, defaults \code{TRUE}}
  \item{buffer}{ integer initial parse buffer size in bytes, adaptively resized as needed: larger buffers can be faster but comsume more memory, default size is 100000L.}
  \item{only_attributes}{ optional logical value, \code{TRUE} means don't retrieve dimension coordinates, only return attribute values; defaults to \code{FALSE}.}
}
}
\keyword{internal}
