\name{scidbdisconnect}
\alias{scidbdisconnect}
\title{
Disconnect from a SciDB database.
}
\description{
SciDB connections are automatically disconnected by the package
in normal practice. The \code{scidbdisconnect} function forces
the current connection to disconnect.
}
\usage{
scidbdisconnect()
}
\value{
NULL is invisibly returned.
}
\author{
B. W. Lewis <blewis@paradigm4.com>
}
