% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,afl-method}
\alias{show,afl-method}
\title{Print a summary of a \code{afl} object}
\usage{
\S4method{show}{afl}(object)
}
\arguments{
\item{object}{\code{afl} object}
}
\value{
printed object summary
}
\description{
Print a summary of a \code{afl} object
}

