% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,scidb-method}
\alias{show,scidb-method}
\title{Print a summary of a \code{scidb} object}
\usage{
\S4method{show}{scidb}(object)
}
\arguments{
\item{object}{a \code{scidb} object}
}
\value{
printed object summary
}
\description{
Print a summary of a \code{scidb} object
}

