% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{store}
\alias{store}
\title{Evaluate an expression to \code{scidb} or \code{scidb} objects}
\usage{
store(db, expr, name, eval = TRUE, gc = TRUE, temp = FALSE)
}
\arguments{
\item{db}{scidb connection object from \code{\link{scidbconnect}}}

\item{expr}{a quoted SciDB expression \code{scidb} object}

\item{name}{(character) optional SciDB array name to store result to}

\item{eval}{\code{FALSE} do not evaluate the expression in SciDB (leave as a view)}

\item{gc}{(logical) optional, when TRUE tie result to R garbage collector}

\item{temp}{(logical, optional), when TRUE store as a SciDB temp array}
}
\description{
Force evaluation of an expression that yields a \code{scidb} or \code{scidb} object,
storing the result to a SciDB array when \code{eval=TRUE}.
}
