\name{summary.scidb}
\alias{summary.scidb}
\title{
Summarize scidb object class.
}
\description{
Summarize scidb object class. This is not a really useful function yet.
}
\usage{
\method{summary}{scidb}(x)
}
\arguments{
  \item{x}{ a scidb array object.
}
}
\author{
B. W. Lewis <blewis@paradigm4.com>
}
