% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl.R
\name{update.afl}
\alias{update.afl}
\title{Update available AFL operators}
\usage{
\method{update}{afl}(.db, .new, .ops)
}
\arguments{
\item{.db}{an \code{afl} object (a SciDB database connection returned from \code{\link{scidbconnect}})}

\item{.new}{a character vector of operator names}

\item{.ops}{an optional three-variable data frame with variables name, signature, help, corresponding
to the operator names, signatures, and help files (from SciDB Doxygen documentation)}
}
\value{
an updated database \code{afl} object
}
\description{
Update available AFL operators
}
\note{
Every operator gets a shallow copy of the db argument; that is, \code{attributes(db[i])$conn} should be the same for every operator index i.
}
\keyword{internal}
