% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{df2scidb}
\alias{df2scidb}
\title{Internal function to upload an R data frame to SciDB}
\usage{
df2scidb(db, X, name = tmpnam(db), types = NULL, chunk_size, gc)
}
\arguments{
\item{db}{scidb database connection}

\item{X}{a data frame}

\item{name}{SciDB array name}

\item{types}{SciDB attribute types}

\item{chunk_size}{optional value passed to the aio_input operator see https://github.com/Paradigm4/accelerated_io_tools}

\item{gc}{set to \code{TRUE} to connect SciDB array to R's garbage collector}
}
\value{
a \code{\link{scidb}} object, or a character schema string if \code{schema_only=TRUE}.
}
\description{
Internal function to upload an R data frame to SciDB
}
\keyword{internal}

