% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dimnames.scidb}
\alias{dimnames.scidb}
\title{Names of array dimensions}
\usage{
\method{dimnames}{scidb}(x)
}
\arguments{
\item{x}{\code{scidb} array object}
}
\value{
a vector of SciDB array dimension names
}
\description{
Names of array dimensions
}

