% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{names.scidb}
\alias{names.scidb}
\title{SciDB dimension and attribute names}
\usage{
\method{names}{scidb}(x)
}
\arguments{
\item{x}{\code{scidb} array object}
}
\value{
Character vector of names
}
\description{
SciDB dimension and attribute names
}

