\name{Ops.scidb}
\alias{Ops.scidb}
\title{
Binary scidb object operations.
}
\description{
Binary scidb object operations.
}
\usage{
\method{Ops}{scidb}(e1, e2)
}
\arguments{
  \item{e1}{
A scalar, vector, or matrix value, or SciDB array reference.
}
  \item{e2}{
A scalar, vector, or matrix value, or SciDB array reference.
}
}
\value{
A scidb or scidbdf object.
}
\author{
B. W. Lewis <blewis@paradigm4.com>
}
