\name{scidb-package}
\alias{scidb-package}
\docType{package}
\title{ SciDB interface package.  }
\description{The scidb package provides basic functions
that move data between R and SciDB and issue SciDB queries.
}
\details{
\tabular{ll}{
Package: \tab scidb\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2012-07-18\cr
License: \tab GPL3\cr
Depends: \tab methods\cr
}
}
\author{
B. W. Lewis <blewis@paradigm4.com>

Maintainer: B. W. Lewis <blewis@paradigm4.com>
}
\keyword{ package }
